19.07 email

+++ src/pages/Assignments.jsx
@@ -19,14 +19,19 @@ const Assignments = () => {
     }, []);
 
     const handleDelete = async (id) => {
-        try {
-            await axios.delete("http://localhost:3001/assignments/" + id);
-            setAssignments(assignments.filter(assignment => assignment.id !== id));
-        } catch (err) {
-            console.log(err);
+        const confirmed = window.confirm("Ви впевнені, що хочете видалити цей проект?");
+        if (confirmed) {
+            try {
+                await axios.delete("http://localhost:3001/assignments/" + id);
+                setAssignments(assignments.filter(assignment => assignment.id !== id));
+            } catch (err) {
+                console.log(err);
+            }
         }
     };
 
+
+    
     const formatDate = (dateString) => {
         const options = { year: 'numeric', month: 'numeric', day: 'numeric' };
         return new Date(dateString).toLocaleDateString(undefined, options);
