+++ src/pages/Users.jsx
@@ -19,11 +19,14 @@ const Users = () => {
     }, []);
 
     const handleDelete = async (id) => {
-        try {
-            await axios.delete("http://localhost:3001/users/" + id);
-            setUsers(users.filter(user => user.id !== id));
-        } catch (err) {
-            console.log(err);
+        const confirmed = window.confirm("Ви впевнені, що хочете видалити цей проект?");
+        if (confirmed) {
+            try {
+                await axios.delete("http://localhost:3001/users/" + id);
+                setUsers(users.filter(user => user.id !== id));
+            } catch (err) {
+                console.log(err);
+            }
         }
     };
